[AIRLFOW-YYY] Move webserver command to separate file
